chore: Download composer installer in seperate stage

+++ docker/docker-entrypoint.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 set -e
 
-if [[ ! -f /root/.composer/composer.json ]]; then
+if [ -f /usr/local/bin/composer ] && [ ! -f /root/.composer/composer.json ]
+then
     echo Installing global composer packages, please wait...
     mv /tmp/composer.json /root/.composer
     composer install \
